Reorganize parameter tests, splitting across unit and integration tests, tweaking test names, uncommenting a few basic ParameterVariable tests that now pass.
